pyproject.toml
[tool.poetry]
name = "bot-kuraku"
version = "1.0.0"
description = "bot-kuraku"
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = { extras = ["fastapi", "httpx"], version = "^2.0.0" }
nb-cli = "^1.0.5"
asyncpg = "^0.27.0"
cachetools = "^5.2.0"
httpx = { version = ">=0.23.0", extras = ["socks"] }
nonebot-adapter-onebot = ">=2.1.5"
nonebot-adapter-qqguild = ">=0.2.2"
nonebot-adapter-kaiheila = "^0.2.6"
nonebot-plugin-escape-url = ">=0.1.0"
nonebot-plugin-gocqhttp-cross-machine-upload-file = ">=0.1.3"
nonebot-plugin-pixivbot = { extras = ["postgresql"], version = ">=2.0.0" }
nonebot-plugin-mahjong-utils = { extras = ["htmlrender"], version = ">=0.3.0" }
nonebot-plugin-mahjong-scoreboard = ">=0.3.0"
nonebot-plugin-majsoul = ">=0.2.0.post4"
nonebot-plugin-nagabus = ">=0.2.0.post1"
nonebot-plugin-access-control = ">=0.5.1"
nonebot_plugin_picstatus = ">=0.3.0"
nonebot-plugin-revoke = ">=0.1.0"
nonebot-plugin-autoreply = ">=0.2.3"
nonebot-plugin-apscheduler = ">=0.2.0"
nonebot-plugin-send-anything-anywhere = ">=0.2.2"

[tool.nonebot]
plugins = ["nonebot_plugin_escape_url", "nonebot_plugin_mahjong_scoreboard", "nonebot_plugin_mahjong_utils", "nonebot_plugin_majsoul",
    "nonebot_plugin_pixivbot",
    "nonebot_plugin_autoreply",
    "nonebot_plugin_access_control", "nonebot_plugin_picstatus", "nonebot_plugin_revoke",
    "nonebot_plugin_nagabus"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
